Updated stoich branch to new organized system
Merged old stoich branch with new organized branch
